+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog
  - Use Item#releaseRecordedStateChanges in RepositoryBasedTransitionHandler to store all state changes
  - Transition#validate also validates post actions
  - Transition#getRequiredPayloadProperties also recognize options of post actions
+ - WorkflowManager#getWorkflowByItem will always return current workflow if given
 
 ### Fixed