+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Changelog
 [Unreleased]
 ------------
 
+[2.1.0] (2020-02-07)
+
 ### Added
 
  - Add method execute to Transition
@@ -72,7 +74,8 @@ Changelog
  - Utilize phpcq/all-tasks.
 
 
-[Unreleased]:  https://github.com/netzmacht/workfow/compare/2.0.1...hotfix/2.0.2
+[Unreleased]:  https://github.com/netzmacht/workfow/compare/master...develop
+[2.1.0]:       https://github.com/netzmacht/workfow/compare/2.0.2...2.1.0
 [2.0.2]:       https://github.com/netzmacht/workfow/compare/2.0.1...2.0.2
 [2.0.1]:       https://github.com/netzmacht/workfow/compare/2.0.0...2.0.1
 [2.0.0]:       https://github.com/netzmacht/workfow/compare/2.0.0-beta1...2.0.0